adicionado a remoção offline dos alunos

+++ app/src/main/java/br/com/alura/agenda/dao/AlunoDAO.java
@@ -18,7 +18,7 @@ import br.com.alura.agenda.modelo.Aluno;
  */
 public class AlunoDAO extends SQLiteOpenHelper {
     public AlunoDAO(Context context) {
-        super(context, "Agenda", null, 5);
+        super(context, "Agenda", null, 6);
     }
 
     @Override
@@ -30,7 +30,8 @@ public class AlunoDAO extends SQLiteOpenHelper {
                 "site TEXT, " +
                 "nota REAL, " +
                 "caminhoFoto TEXT," +
-                "sincronizado INT DEFAULT 0);";
+                "sincronizado INT DEFAULT 0," +
+                "desativado INT DEFAULT 0);";
         db.execSQL(sql);
     }
 
@@ -81,6 +82,11 @@ public class AlunoDAO extends SQLiteOpenHelper {
             case 4:
                 String adicionaCampoSincronizado = "ALTER TABLE Alunos ADD COLUMN sincronizado INT DEFAULT 0";
                 db.execSQL(adicionaCampoSincronizado);
+
+            case 5:
+                String adicionaCampoDesativado =
+                        "ALTER TABLE Alunos ADD COLUMN desativado INT DEFAULT 0";
+                db.execSQL(adicionaCampoDesativado);
         }
 
     }
@@ -113,17 +119,17 @@ public class AlunoDAO extends SQLiteOpenHelper {
         dados.put("nota", aluno.getNota());
         dados.put("caminhoFoto", aluno.getCaminhoFoto());
         dados.put("sincronizado", aluno.getSincronizado());
+        dados.put("desativado", aluno.getDesativado());
         return dados;
     }
 
     public List<Aluno> buscaAlunos() {
-        String sql = "SELECT * FROM Alunos;";
+        String sql = "SELECT * FROM Alunos WHERE desativado = 0;";
         SQLiteDatabase db = getReadableDatabase();
         Cursor c = db.rawQuery(sql, null);
 
         List<Aluno> alunos = populaAlunos(c);
         c.close();
-
         return alunos;
     }
 
@@ -140,6 +146,7 @@ public class AlunoDAO extends SQLiteOpenHelper {
             aluno.setNota(c.getDouble(c.getColumnIndex("nota")));
             aluno.setCaminhoFoto(c.getString(c.getColumnIndex("caminhoFoto")));
             aluno.setSincronizado(c.getInt(c.getColumnIndex("sincronizado")));
+            aluno.setDesativado(c.getColumnIndex("desativado"));
             alunos.add(aluno);
         }
         return alunos;
@@ -149,7 +156,12 @@ public class AlunoDAO extends SQLiteOpenHelper {
         SQLiteDatabase db = getWritableDatabase();
 
         String[] params = {aluno.getId().toString()};
+        if(aluno.estaDesativado()){
         db.delete("Alunos", "id = ?", params);
+        }else{
+            aluno.desativa();
+            altera(aluno);
+        }
     }
 
     public void altera(Aluno aluno) {
@@ -158,7 +170,9 @@ public class AlunoDAO extends SQLiteOpenHelper {
         ContentValues dados = pegaDadosDoAluno(aluno);
 
         String[] params = {aluno.getId().toString()};
+
         db.update("Alunos", dados, "id = ?", params);
+
     }
 
     public boolean ehAluno(String telefone) {
